Decoupled logic parts from interface parts
