+++ Randomizer/BaseClasses/BaseTargetOp.py
@@ -1,5 +1,7 @@
+from typing import IO
+
 
 class BaseTargetOp(object):
-    def __init__(self, file, params = None):
+    def __init__(self, file: IO, params = None):
         self.params = params
         self.file = file
