+++ Randomizer/BaseClasses/PatchBase.py
@@ -1,5 +1,6 @@
 from BaseClasses.BaseTargetOp import BaseTargetOp
 
+
 class PatchBase(BaseTargetOp):
 	def __init__(self, file, params):
 		super().__init__(file, params)
